+++ pallets/subtensor/src/extensions/mod.rs
@@ -1,0 +1,3 @@
+mod subtensor;
+
+pub use subtensor::*;
